refactor: cleaned up some filepaths
