+++ src/core/server/models/action/moderation/comment.ts
@@ -8,11 +8,11 @@ import {
   ConnectionInput,
   getPageInfo,
   nodesToEdges,
-} from "talk-server/models/connection";
+} from "talk-server/models/helpers/connection";
 import Query, {
   createConnectionOrderVariants,
   createIndexFactory,
-} from "talk-server/models/query";
+} from "talk-server/models/helpers/query";
 import { TenantResource } from "talk-server/models/tenant";
 
 function collection(db: Db) {
@@ -21,18 +21,18 @@ function collection(db: Db) {
   );
 }
 
-const createCommentModerationActionConnectionOrderVariants = createConnectionOrderVariants<
-  Readonly<CommentModerationAction>
->([{ createdAt: -1 }]);
-
 export async function createCommentModerationActionIndexes(mongo: Db) {
   const createIndex = createIndexFactory(collection(mongo));
 
   // UNIQUE { id }
   await createIndex({ tenantID: 1, id: 1 }, { unique: true });
 
+  const createVariants = createConnectionOrderVariants<
+    Readonly<CommentModerationAction>
+  >([{ createdAt: -1 }]);
+
   // { moderatorID, ...connectionParams }
-  await createCommentModerationActionConnectionOrderVariants(createIndex, {
+  await createVariants(createIndex, {
     moderatorID: 1,
   });
 }
