+++ src/core/server/models/helpers/connection.ts
@@ -1,6 +1,6 @@
 import { merge } from "lodash";
 
-import { FilterQuery } from "talk-server/models/query";
+import { FilterQuery } from "talk-server/models/helpers/query";
 
 export type Cursor = Date | number | string | null;
